Fix file update on classes page

+++ src/api/classes.js
@@ -34,15 +34,18 @@ export const classes = {
             },
         );
     },
-    update: async ({id, newDoc, fileName, callback}) => {
-        const {file} = newDoc;
+    update: async ({id, newDoc, callback}) => {
+        const {file, fileName, image} = newDoc;
+        console.log(newDoc)
+
         if (!file) {
-            if (fileName) {
+            if (image && fileName) {
                 await firebaseStorage
                     .ref()
                     .child(`classes/${fileName}`)
                     .delete();
-            }
+                delete newDoc.fileName;
+            } 
             await classesCol.doc(id).set(newDoc);
             return callback({...newDoc, id});
         } else {
@@ -52,6 +55,7 @@ export const classes = {
                     .child(`classes/${fileName}`)
                     .delete();
             }
+
             const fileExtension = newDoc.file.name.split('.').slice(-1)[0];
             const name = `${uuid()}.${fileExtension}`;
             const task = firebaseStorage
